+++ cmd/nibid/cmd/testnet.go
@@ -45,7 +45,9 @@ var (
 )
 
 // get cmd to initialize all files for tendermint testnet and application
-func testnetCmd(mbm module.BasicManager, genBalIterator banktypes.GenesisBalancesIterator) *cobra.Command {
+func testnetCmd(
+	mbm module.BasicManager, genBalIterator banktypes.GenesisBalancesIterator,
+) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "testnet",
 		Short: "Initialize files for a simapp testnet",