+++ x/evm/tx.go
@@ -1,0 +1,361 @@
+// Copyright (c) 2023-2024 Nibi, Inc.
+package evm
+
+import (
+	"math"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
+	gethmath "github.com/ethereum/go-ethereum/common/math"
+	gethcore "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/core/vm"
+
+	errorsmod "cosmossdk.io/errors"
+	sdkmath "cosmossdk.io/math"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"github.com/NibiruChain/nibiru/eth"
+)
+
+// EvmTxArgs encapsulates all possible params to create all EVM txs types.
+// This includes LegacyTx, DynamicFeeTx and AccessListTx
+type EvmTxArgs struct { //revive:disable-line:exported
+	Nonce     uint64
+	GasLimit  uint64
+	Input     []byte
+	GasFeeCap *big.Int
+	GasPrice  *big.Int
+	ChainID   *big.Int
+	Amount    *big.Int
+	GasTipCap *big.Int
+	To        *common.Address
+	Accesses  *gethcore.AccessList
+}
+
+// DefaultPriorityReduction is the default amount of price values required for 1 unit of priority.
+// Because priority is `int64` while price is `big.Int`, it's necessary to scale down the range to keep it more pratical.
+// The default value is the same as the `sdk.DefaultPowerReduction`.
+var DefaultPriorityReduction = sdk.DefaultPowerReduction
+
+// GetTxPriority returns the priority of a given Ethereum tx. It relies of the
+// priority reduction global variable to calculate the tx priority given the tx
+// tip price:
+//
+//	tx_priority = tip_price / priority_reduction
+func GetTxPriority(txData TxData, baseFee *big.Int) (priority int64) {
+	// calculate priority based on effective gas price
+	tipPrice := txData.EffectiveGasPrice(baseFee)
+	// if london hardfork is not enabled, tipPrice is the gasPrice
+	if baseFee != nil {
+		tipPrice = new(big.Int).Sub(tipPrice, baseFee)
+	}
+
+	priority = math.MaxInt64
+	priorityBig := new(big.Int).Quo(tipPrice, DefaultPriorityReduction.BigInt())
+
+	// safety check
+	if priorityBig.IsInt64() {
+		priority = priorityBig.Int64()
+	}
+
+	return priority
+}
+
+// Failed returns if the contract execution failed in vm errors
+func (m *MsgEthereumTxResponse) Failed() bool {
+	return len(m.VmError) > 0
+}
+
+// Return is a helper function to help caller distinguish between revert reason
+// and function return. Return returns the data after execution if no error occurs.
+func (m *MsgEthereumTxResponse) Return() []byte {
+	if m.Failed() {
+		return nil
+	}
+	return common.CopyBytes(m.Ret)
+}
+
+// Revert returns the concrete revert reason if the execution is aborted by `REVERT`
+// opcode. Note the reason can be nil if no data supplied with revert opcode.
+func (m *MsgEthereumTxResponse) Revert() []byte {
+	if m.VmError != vm.ErrExecutionReverted.Error() {
+		return nil
+	}
+	return common.CopyBytes(m.Ret)
+}
+
+func NewDynamicFeeTx(tx *gethcore.Transaction) (*DynamicFeeTx, error) {
+	txData := &DynamicFeeTx{
+		Nonce:    tx.Nonce(),
+		Data:     tx.Data(),
+		GasLimit: tx.Gas(),
+	}
+
+	v, r, s := tx.RawSignatureValues()
+	if to := tx.To(); to != nil {
+		txData.To = to.Hex()
+	}
+
+	if tx.Value() != nil {
+		amountInt, err := eth.SafeNewIntFromBigInt(tx.Value())
+		if err != nil {
+			return nil, err
+		}
+		txData.Amount = &amountInt
+	}
+
+	if tx.GasFeeCap() != nil {
+		gasFeeCapInt, err := eth.SafeNewIntFromBigInt(tx.GasFeeCap())
+		if err != nil {
+			return nil, err
+		}
+		txData.GasFeeCap = &gasFeeCapInt
+	}
+
+	if tx.GasTipCap() != nil {
+		gasTipCapInt, err := eth.SafeNewIntFromBigInt(tx.GasTipCap())
+		if err != nil {
+			return nil, err
+		}
+		txData.GasTipCap = &gasTipCapInt
+	}
+
+	if tx.AccessList() != nil {
+		al := tx.AccessList()
+		txData.Accesses = NewAccessList(&al)
+	}
+
+	txData.SetSignatureValues(tx.ChainId(), v, r, s)
+	return txData, nil
+}
+
+// TxType returns the tx type
+func (tx *DynamicFeeTx) TxType() uint8 {
+	return gethcore.DynamicFeeTxType
+}
+
+// Copy returns an instance with the same field values
+func (tx *DynamicFeeTx) Copy() TxData {
+	return &DynamicFeeTx{
+		ChainID:   tx.ChainID,
+		Nonce:     tx.Nonce,
+		GasTipCap: tx.GasTipCap,
+		GasFeeCap: tx.GasFeeCap,
+		GasLimit:  tx.GasLimit,
+		To:        tx.To,
+		Amount:    tx.Amount,
+		Data:      common.CopyBytes(tx.Data),
+		Accesses:  tx.Accesses,
+		V:         common.CopyBytes(tx.V),
+		R:         common.CopyBytes(tx.R),
+		S:         common.CopyBytes(tx.S),
+	}
+}
+
+// GetChainID returns the chain id field from the DynamicFeeTx
+func (tx *DynamicFeeTx) GetChainID() *big.Int {
+	if tx.ChainID == nil {
+		return nil
+	}
+
+	return tx.ChainID.BigInt()
+}
+
+// GetAccessList returns the AccessList field.
+func (tx *DynamicFeeTx) GetAccessList() gethcore.AccessList {
+	if tx.Accesses == nil {
+		return nil
+	}
+	return *tx.Accesses.ToEthAccessList()
+}
+
+// GetData returns the a copy of the input data bytes.
+func (tx *DynamicFeeTx) GetData() []byte {
+	return common.CopyBytes(tx.Data)
+}
+
+// GetGas returns the gas limit.
+func (tx *DynamicFeeTx) GetGas() uint64 {
+	return tx.GasLimit
+}
+
+// GetGasPrice returns the gas fee cap field.
+func (tx *DynamicFeeTx) GetGasPrice() *big.Int {
+	return tx.GetGasFeeCap()
+}
+
+// GetGasTipCap returns the gas tip cap field.
+func (tx *DynamicFeeTx) GetGasTipCap() *big.Int {
+	if tx.GasTipCap == nil {
+		return nil
+	}
+	return tx.GasTipCap.BigInt()
+}
+
+// GetGasFeeCap returns the gas fee cap field.
+func (tx *DynamicFeeTx) GetGasFeeCap() *big.Int {
+	if tx.GasFeeCap == nil {
+		return nil
+	}
+	return tx.GasFeeCap.BigInt()
+}
+
+// GetValue returns the tx amount.
+func (tx *DynamicFeeTx) GetValue() *big.Int {
+	if tx.Amount == nil {
+		return nil
+	}
+
+	return tx.Amount.BigInt()
+}
+
+// GetNonce returns the account sequence for the transaction.
+func (tx *DynamicFeeTx) GetNonce() uint64 { return tx.Nonce }
+
+// GetTo returns the pointer to the recipient address.
+func (tx *DynamicFeeTx) GetTo() *common.Address {
+	if tx.To == "" {
+		return nil
+	}
+	to := common.HexToAddress(tx.To)
+	return &to
+}
+
+// AsEthereumData returns an DynamicFeeTx transaction tx from the proto-formatted
+// TxData defined on the Cosmos EVM.
+func (tx *DynamicFeeTx) AsEthereumData() gethcore.TxData {
+	v, r, s := tx.GetRawSignatureValues()
+	return &gethcore.DynamicFeeTx{
+		ChainID:    tx.GetChainID(),
+		Nonce:      tx.GetNonce(),
+		GasTipCap:  tx.GetGasTipCap(),
+		GasFeeCap:  tx.GetGasFeeCap(),
+		Gas:        tx.GetGas(),
+		To:         tx.GetTo(),
+		Value:      tx.GetValue(),
+		Data:       tx.GetData(),
+		AccessList: tx.GetAccessList(),
+		V:          v,
+		R:          r,
+		S:          s,
+	}
+}
+
+// GetRawSignatureValues returns the V, R, S signature values of the transaction.
+// The return values should not be modified by the caller.
+func (tx *DynamicFeeTx) GetRawSignatureValues() (v, r, s *big.Int) {
+	return rawSignatureValues(tx.V, tx.R, tx.S)
+}
+
+// SetSignatureValues sets the signature values to the transaction.
+func (tx *DynamicFeeTx) SetSignatureValues(chainID, v, r, s *big.Int) {
+	if v != nil {
+		tx.V = v.Bytes()
+	}
+	if r != nil {
+		tx.R = r.Bytes()
+	}
+	if s != nil {
+		tx.S = s.Bytes()
+	}
+	if chainID != nil {
+		chainIDInt := sdkmath.NewIntFromBigInt(chainID)
+		tx.ChainID = &chainIDInt
+	}
+}
+
+// Validate performs a stateless validation of the tx fields.
+func (tx DynamicFeeTx) Validate() error {
+	if tx.GasTipCap == nil {
+		return errorsmod.Wrap(ErrInvalidGasCap, "gas tip cap cannot nil")
+	}
+
+	if tx.GasFeeCap == nil {
+		return errorsmod.Wrap(ErrInvalidGasCap, "gas fee cap cannot nil")
+	}
+
+	if tx.GasTipCap.IsNegative() {
+		return errorsmod.Wrapf(ErrInvalidGasCap, "gas tip cap cannot be negative %s", tx.GasTipCap)
+	}
+
+	if tx.GasFeeCap.IsNegative() {
+		return errorsmod.Wrapf(ErrInvalidGasCap, "gas fee cap cannot be negative %s", tx.GasFeeCap)
+	}
+
+	if !eth.IsValidInt256(tx.GetGasTipCap()) {
+		return errorsmod.Wrap(ErrInvalidGasCap, "out of bound")
+	}
+
+	if !eth.IsValidInt256(tx.GetGasFeeCap()) {
+		return errorsmod.Wrap(ErrInvalidGasCap, "out of bound")
+	}
+
+	if tx.GasFeeCap.LT(*tx.GasTipCap) {
+		return errorsmod.Wrapf(
+			ErrInvalidGasCap, "max priority fee per gas higher than max fee per gas (%s > %s)",
+			tx.GasTipCap, tx.GasFeeCap,
+		)
+	}
+
+	if !eth.IsValidInt256(tx.Fee()) {
+		return errorsmod.Wrap(ErrInvalidGasFee, "out of bound")
+	}
+
+	amount := tx.GetValue()
+	// Amount can be 0
+	if amount != nil && amount.Sign() == -1 {
+		return errorsmod.Wrapf(ErrInvalidAmount, "amount cannot be negative %s", amount)
+	}
+	if !eth.IsValidInt256(amount) {
+		return errorsmod.Wrap(ErrInvalidAmount, "out of bound")
+	}
+
+	if tx.To != "" {
+		if err := eth.ValidateAddress(tx.To); err != nil {
+			return errorsmod.Wrap(err, "invalid to address")
+		}
+	}
+
+	chainID := tx.GetChainID()
+
+	if chainID == nil {
+		return errorsmod.Wrap(
+			errortypes.ErrInvalidChainID,
+			"chain ID must be present on AccessList txs",
+		)
+	}
+
+	return nil
+}
+
+// Fee returns gasprice * gaslimit.
+func (tx DynamicFeeTx) Fee() *big.Int {
+	return fee(tx.GetGasFeeCap(), tx.GasLimit)
+}
+
+// Cost returns amount + gasprice * gaslimit.
+func (tx DynamicFeeTx) Cost() *big.Int {
+	return cost(tx.Fee(), tx.GetValue())
+}
+
+// EffectiveGasPrice computes the effective gas price based on eip-1559 rules
+// `effectiveGasPrice = min(baseFee + tipCap, feeCap)`
+func EffectiveGasPrice(baseFee, feeCap, tipCap *big.Int) *big.Int {
+	return gethmath.BigMin(new(big.Int).Add(tipCap, baseFee), feeCap)
+}
+
+// EffectiveGasPrice returns the effective gas price
+func (tx *DynamicFeeTx) EffectiveGasPrice(baseFee *big.Int) *big.Int {
+	return EffectiveGasPrice(baseFee, tx.GasFeeCap.BigInt(), tx.GasTipCap.BigInt())
+}
+
+// EffectiveFee returns effective_gasprice * gaslimit.
+func (tx DynamicFeeTx) EffectiveFee(baseFee *big.Int) *big.Int {
+	return fee(tx.EffectiveGasPrice(baseFee), tx.GasLimit)
+}
+
+// EffectiveCost returns amount + effective_gasprice * gaslimit.
+func (tx DynamicFeeTx) EffectiveCost(baseFee *big.Int) *big.Int {
+	return cost(tx.EffectiveFee(baseFee), tx.GetValue())
+}